refactor: setting up semantic folder structure
