+++ .babelrc
@@ -1,4 +1,9 @@
 {
 	"presets": ["next/babel"],
-	"plugins": [["styled-components", { "ssr": true }]]
+	"plugins": [
+		[
+			"styled-components",
+			{ "ssr": true, "displayName": true, "preprocess": false }
+		]
+	]
 }
