+++ core/graphql/queries.ts
