+++ core/utils/slice.ts
