+++ package.json
@@ -12,43 +12,43 @@
 		"fix": "eslint --fix ."
 	},
 	"dependencies": {
-		"date-fns": "^2.16.1",
-		"disqus-react": "^1.0.11",
-		"framer-motion": "^3.3.0",
-		"graphql": "^15.5.0",
-		"graphql-request": "^3.4.0",
-		"next": "^11.0.1",
+		"date-fns": "^2.23.0",
+		"disqus-react": "^1.1.1",
+		"framer-motion": "^3.10.6",
+		"graphql": "^15.5.1",
+		"graphql-request": "^3.5.0",
+		"next": "^11.1.0",
 		"next-themes": "0.0.10",
 		"nprogress": "^0.2.0",
-		"prism-react-renderer": "^1.2.0",
+		"prism-react-renderer": "^1.2.1",
 		"react": "17.0.1",
 		"react-dom": "17.0.1",
-		"react-icons": "^4.1.0",
+		"react-icons": "^4.2.0",
 		"react-loader-spinner": "^4.0.0",
 		"react-markdown": "^5.0.3",
-		"react-syntax-highlighter": "^15.4.3",
-		"reading-time": "^1.3.0",
-		"styled-components": "^5.2.1",
+		"react-syntax-highlighter": "^15.4.4",
+		"reading-time": "^1.4.0",
+		"styled-components": "^5.3.0",
 		"terminal-in-react": "^4.3.1"
 	},
 	"devDependencies": {
 		"@commitlint/cli": "^13.1.0",
 		"@commitlint/config-conventional": "^13.1.0",
-		"@types/node": "^14.14.22",
-		"@types/react": "^17.0.0",
-		"@types/react-dom": "^17.0.0",
-		"@types/styled-components": "^5.1.7",
-		"@typescript-eslint/eslint-plugin": "^4.28.5",
-		"@typescript-eslint/parser": "^4.28.5",
-		"babel-plugin-styled-components": "^1.12.0",
-		"eslint": "^7.31.0",
-		"eslint-config-next": "^11.0.1",
+		"@types/node": "^14.17.11",
+		"@types/react": "^17.0.19",
+		"@types/react-dom": "^17.0.9",
+		"@types/styled-components": "^5.1.12",
+		"@typescript-eslint/eslint-plugin": "^4.29.2",
+		"@typescript-eslint/parser": "^4.29.2",
+		"babel-plugin-styled-components": "^1.13.2",
+		"eslint": "^7.32.0",
+		"eslint-config-next": "^11.1.0",
 		"eslint-plugin-react": "^7.24.0",
 		"husky": "^4.3.8",
-		"lint-staged": "^11.1.1",
+		"lint-staged": "^11.1.2",
 		"prettier": "^2.3.2",
-		"typescript": "^4.1.3",
-		"webpack": "^5.47.0"
+		"typescript": "^4.3.5",
+		"webpack": "^5.51.1"
 	},
 	"husky": {
 		"hooks": {
