+++ pages/_app.tsx
@@ -4,7 +4,7 @@ import { createGlobalStyle } from 'styled-components';
 import { ThemeProvider } from 'next-themes';
 import { AnimatePresence } from 'framer-motion';
 import NProgress from 'nprogress';
-import '../public/asset/css/nprogress.min.css';
+import '@public/asset/css/nprogress.min.css';
 
 NProgress.configure({
 	minimum: 0.3,
