+++ pages/about.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Head from 'next/head';
-import Footer from '../components/Footer';
-import AboutContent from '../components/About';
-import Navbar from '../components/Navbar';
+import Footer from '@components/Footer';
+import AboutContent from '@components/About';
+import Navbar from '@components/Navbar';
 
 const About: React.FC = () => {
 	return (
